Added Feature:
Progress bar now working for Osias Encryption
Save File (Key and Output) now adapts to save folder
Output and Key can now be deleted since disposed function is applies

+++ DIP Algorithm/OSIASEncryption.cs
@@ -130,7 +130,7 @@ namespace DIP_Algorithm
                     }
                     if (source.Position >= source.Length)
                         break;
-                    Percentage = source.Position / source.Length;
+                    Percentage = (float)source.Position / (float)source.Length;
                 }
             }
             this.Output = new EncryptionMeta(dataBitmap, keyBitmap);
